HailoRT-Ui/requirements.txt
# Hailo-H8 Control Panel Dependencies

# Qt5 UI Framework
PyQt5>=5.15.0

# Core dependencies
numpy>=1.21.0
Pillow>=9.0.0
PyYAML>=6.0

# Video/Image processing
opencv-python>=4.5.0

# PyTorch (for model conversion)
torch>=1.9.0
torchvision>=0.10.0

# ONNX (for model verification)
onnx>=1.12.0
onnxruntime>=1.12.0

# YOLO support (optional)
# ultralytics>=8.0.0

# Hailo SDK (install separately from Hailo website)
# hailort
# hailo_sdk_client
